server/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
Authlib==1.6.9
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
fastapi==0.135.1
h11==0.16.0
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
pycparser==3.0
python-multipart==0.0.9
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.15.3
python-dotenv==1.2.2
python-jose[cryptography]==3.3.0
mongomock==4.3.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
websockets==15.0.1
pywinpty==3.0.3; sys_platform == 'win32'
